refactor/organize
