+++ Libraries/AssemblyLib/Patching/IModulePatch.cs
